# pe: cache DataAfterSection for performance