@stevemk14ebr 2023-07-21
Thanks for the performance improvements Willi! This has been a long time goal, I'm impressed by the results so far. I will review and merge these as soon as I can