@williballenthin 2023-08-02
I also had an implementation that copied all the section data into a single contiguous buffer and passed this around. It was approximately as fast without the additional copies. However, I was working against v2.3 due to not having yara and v2.4 working, and there were merge conflicts when updating to v2.4, so all I have is this PR.

sidebar: I would recommend removing the yara dependency, but we can discuss this elsewhere.