@stevemk14ebr 2023-08-15
I noticed some issues with this PR. The new `dataAfterSection` routine doesn't actually get all the data after a given section, it only returns the data of the particular found section. We also need to do this optimization for the other OS's